requirements.txt
pillow==10.2.0
pillow-heif==0.14.0
